\name{canceR}
\alias{canceR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{canceR}
\description{A Graphical User Interface for accessing and modeling the MSKCC Cancer Genomics Data. }
\details{
The package is user friendly interface based on the cgdsr and other modeling packages (see depends) to explore, compare, and analyse all available Cancer Data (Clinical data, Gene Mutation, Gene Methylation, Gene Expression, Protein Phosphorylation, Copy Number Alteration) hosted by the Computational Biology Center at Memorial-Sloan-Kettering Cancer Center (MSKCC). 
}
\usage{
canceR()
}


\references{
Karim Mezhoud (in press) canceR: A Graphical User Interface for accessing and modeling the MSKCC Cancer Genomics Data.
}
\author{
Karim Mezhoud
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ cgdsr }
\keyword{ cancer }
\keyword{ GUI }

\value{
Value
}
%\examples{
%install.packages("canceR")
%library("canceR")
%canceR()
%}

\examples{


canceR()

}