\name{canceR_Vignette}
\alias{canceR_Vignette}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
canceR_Vignette
}
\description{
CanceR Manual
}
\usage{
canceR_Vignette()
}
%- maybe also 'usage' for other objects documented here.

\references{
In press
}
\author{
Karim mezhoud
}
\value{
Value
}
\examples{


canceR_vignette()

}