\name{dialogPlotOption_SkinCor}
\alias{dialogPlotOption_SkinCor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
dialogPlotOption_SkinCor
}
\description{
Specifying Skin and Correlation Options.
}
\usage{
dialogPlotOption_SkinCor()
}
%- maybe also 'usage' for other objects documented here.

\seealso{
cgdsr package
}
\value{
Value
}
\examples{


dialogPlotOption_skinCor()

}