\name{getCasesGenProfs}
\alias{getCasesGenProfs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getCasesGenProfs
}
\description{
Get Genetic profiles from cbioportal.
}
\usage{
getCasesGenProfs()
}
%- maybe also 'usage' for other objects documented here.

\value{
Value
}
\examples{


getCasesGenProfs()

}