\name{getGeneListExample}
\alias{getGeneListExample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getGeneListExample
}
\description{
get example of gene list.
}
\usage{
getGeneListExample()
}
%- maybe also 'usage' for other objects documented here.
\value{
Value
}
\examples{


getGeneListExample()

}
