\name{getGenesClassifier}
\alias{getGenesClassifier}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getGenesClassifier
}
\description{
refers to geNetClassifier package.
}
\usage{
getGenesClassifier()
}
%- maybe also 'usage' for other objects documented here.

\seealso{
geNetClassifier
}

\value{
Value
}
\examples{


getGenesClassifier()

}