\name{getGenesTree_MultipleCases}
\alias{getGenesTree_MultipleCases}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getGenesTree_MultipleCases
}
\description{
Get  successively trees of genes list for multiple cases
}
\value{
Value
}
\examples{

getGenesTree_MultipleCases()

}

