\name{getMetDataMultipleGenes}
\alias{getMetDataMultipleGenes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getMetDataMultipleGenes
}
\description{
Get methylation data for gene list.
}
\usage{
getMetDataMultipleGenes()
}

\value{
Value
}
\examples{

getMetDataMultipleGenes()

}
