\name{getProfilesDataSingleGene}
\alias{getProfilesDataSingleGene}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getProfilesDataSingleGene
}
\description{
Get profile data for single gene.
}
\usage{
getProfilesDataSingleGene()
}
\value{
Value
}
\examples{

getProfileDataSingleGene()

}