\name{getSummaryGSEA}
\alias{getSummaryGSEA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getSummaryGSEA
}
\description{
get summary results of GSEA
}
\usage{
getSummaryGSEA()
}
%- maybe also 'usage' for other objects documented here.


\references{

Karim Mezhoud (in press) canceR: A Graphical User Interface for accessing and modeling the MSKCC Cancer Genomics Data.
}
\author{
Karim Mezhoud
}

\value{
Value
}

\examples{


getSummaryGSEA()

}