\name{Match_GeneList_MSigDB}
\alias{Match_GeneList_MSigDB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Match_GeneList_MSigDB
}
\description{
Search MSigDb that overlap gene list
}
\usage{
Match_GeneList_MSigDB()
}
%- maybe also 'usage' for other objects documented here.

\value{
Value
}
\examples{
Match_GeneList_MSigDB()
}