\name{plot_1Gene_2GenProfs}
\alias{plot_1Gene_2GenProfs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plot_1Gene_2GenProfs
}
\description{
Tcltk plotting of 1 gene depending two genetic profiles (expression vs mutation or exprssion vs methylation)
}
\usage{
plot_1Gene_2GenProfs()
}
%- maybe also 'usage' for other objects documented here.
\value{
Value
}
\examples{
plot_1Gene_2GenProfs()
}