\name{plot_2Genes_1GenProf}
\alias{plot_2Genes_1GenProf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plot_2Genes_1GenProf
}
\description{
Plotting the correlation between two genes.
}
\usage{
plot_2Genes_1GenProf()
}
%- maybe also 'usage' for other objects documented here.
\value{
Value
}
\examples{
plot_2Genes_1GenProf()
}