\name{testCheckedCaseGenProf}
\alias{testCheckedCaseGenProf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Testing checked appropriate Cases for appropriate Genetic profiles.
}
\description{
Testing checked appropriate Cases for appropriate Genetic profiles.
}
\usage{
testCheckedCaseGenProf()
}


\references{
in press
}
\author{
Karim mezhoud
}

\value{
Value
}
\examples{
testCheckedCaseGenProf()
}